+++ apps/note/src/cli.mjs
@@ -9,6 +9,7 @@ import catCommand from './commands/cat.command.mjs';
 import deleteCommand from './commands/delete.command.mjs';
 import editCommand from './commands/edit.command.mjs';
 import migrateCommand from './commands/migrate.command.mjs';
+import setCommand from './commands/set.command.mjs';
 
 new Command(pkg.name)
   .description(pkg.description)
@@ -20,4 +21,5 @@ new Command(pkg.name)
   .addCommand(listCommand(pkg.name))
   .addCommand(migrateCommand)
   .addCommand(renameCommand(pkg.name))
+  .addCommand(setCommand)
   .parse();