+++ apps/note/src/lib/read-stdin.mjs
@@ -1,0 +1,31 @@
+/**
+ * @param {number} timeout
+ * @returns {Promise<Buffer[]>}
+ */
+export default async function (timeout) {
+  /** @type {Buffer[]} */
+  const chunks = [];
+
+  let timeoutInstance = setTimeout(() => {
+    console.error('Timed out awaiting input from stdin');
+    process.exit(1);
+  }, timeout);
+
+  const stdinRead = new Promise((resolve) => {
+    process.stdin.on('end', resolve);
+  });
+
+  process.stdin.on('data', (data) => {
+    clearTimeout(timeoutInstance);
+    chunks.push(data);
+  });
+
+  await stdinRead;
+
+  if (!chunks.some((buf) => buf.length)) {
+    console.error('No input from stdin');
+    process.exit(1);
+  }
+
+  return chunks;
+}